requirements.txt
arabic_reshaper==3.0.0
Pillow==10.3.0
streamlit==1.42.2
~treamlit==1.36.0
